refactor: centralize type guards
