+++ .changeset/shared-type-guards.md
@@ -1,0 +1,6 @@
+---
+'@lapidist/design-lint': patch
+---
+
+refactor: centralize type guards
+
