+++ src/cli/tokens.ts
@@ -3,6 +3,7 @@ import path from 'path';
 import { loadConfig } from '../config/loader.js';
 import { getFlattenedTokens } from '../core/token-utils.js';
 import type { DesignTokens } from '../core/types.js';
+import { isDesignTokens, isThemeRecord } from '../utils/type-guards.js';
 
 interface TokensCommandOptions {
   theme?: string;
@@ -41,41 +42,4 @@ function toThemeRecord(tokens: unknown): Record<string, DesignTokens> {
     return { default: tokens };
   }
   return {};
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null;
-}
-
-function isDesignTokens(value: unknown): value is DesignTokens {
-  return isRecord(value);
-}
-
-function isThemeRecord(val: unknown): val is Record<string, DesignTokens> {
-  if (!isRecord(val)) return false;
-  const entries = Object.entries(val).filter(([k]) => !k.startsWith('$'));
-  if (entries.length === 0) return false;
-  if (entries.length === 1) {
-    const [, theme] = entries[0];
-    if (!isRecord(theme)) return false;
-    const children = Object.entries(theme)
-      .filter(([k]) => !k.startsWith('$'))
-      .map(([, v]) => v);
-    const allTokens = children.every(
-      (child) => isRecord(child) && ('$value' in child || 'value' in child),
-    );
-    return !allTokens;
-  }
-  let shared: string[] | null = null;
-  for (const [, theme] of entries) {
-    if (!isRecord(theme)) return false;
-    const keys = Object.keys(theme).filter((k) => !k.startsWith('$'));
-    if (shared === null) {
-      shared = keys;
-    } else {
-      shared = shared.filter((k) => keys.includes(k));
-      if (shared.length === 0) return false;
-    }
-  }
-  return true;
 }
